Created unit tests for Wheel Class
created unit tests for Wheel.cpp inside of tests/WheelTests.cpp
updated physics logic inside of Wheel.cpp

+++ core_engine/tests/FuelTankTests.cpp
@@ -8,8 +8,8 @@ class FuelPercentageTest : public ::testing::TestWithParam<std::pair<int,float>>
 
 
 TEST_P(FuelConsumptionTest, ConsumeFuel){
-    int consumedFuel = GetParam().first;
-    int expectedLevel = GetParam().second;
+    float consumedFuel = GetParam().first;
+    float expectedLevel = GetParam().second;
 
     FuelTank fuelTank(100);
     fuelTank.consume(consumedFuel);
@@ -20,12 +20,12 @@ INSTANTIATE_TEST_SUITE_P(
     FuelTankConsumptionTests,
     FuelConsumptionTest,
     ::testing::Values(
-        std::make_pair(90, 10),
-        std::make_pair(100, 0),
-        std::make_pair(150, 0),
-        std::make_pair(80, 20),
-        std::make_pair(0, 100),
-        std::make_pair(-25, 100)
+        std::make_pair(90.0f, 10.0f),
+        std::make_pair(100.0f, 0.0f),
+        std::make_pair(150.0f, 0.0f),
+        std::make_pair(80.0f, 20.0f),
+        std::make_pair(0.0f, 100.0f),
+        std::make_pair(-25.0f, 100.0f)
     )
 );
 
@@ -42,20 +42,20 @@ INSTANTIATE_TEST_SUITE_P(
     FuelTankLowTests,
     FuelLowTest,
     ::testing::Values(
-        std::make_pair(90, true),
-        std::make_pair(100, true),
-        std::make_pair(80, false),
-        std::make_pair(0, false),
-        std::make_pair(-25, false)
+        std::make_pair(90.0f, true),
+        std::make_pair(100.0f, true),
+        std::make_pair(80.0f, false),
+        std::make_pair(0.0f, false),
+        std::make_pair(-25.0f, false)
     )
 );
 
 TEST_P(FuelRefuelTest, RefuelFuel){
-    int refueledAmount = GetParam().first;
-    int expectedLevel = GetParam().second;
+    float refueledAmount = GetParam().first;
+    float expectedLevel = GetParam().second;
 
     FuelTank fuelTank(100);
-    fuelTank.consume(50); // Start at 50% fuel
+    fuelTank.consume(50.0f); // Start at 50% fuel
     fuelTank.refuel(refueledAmount);
     EXPECT_EQ(fuelTank.getCurrentLevel(), expectedLevel);
 }
@@ -64,16 +64,16 @@ INSTANTIATE_TEST_SUITE_P(
     FuelTankRefuelTests,
     FuelRefuelTest,
     ::testing::Values(
-        std::make_pair(20, 70),
-        std::make_pair(50, 100),
-        std::make_pair(60, 100),
-        std::make_pair(0, 50),
-        std::make_pair(-25, 50)
+        std::make_pair(20.0f, 70.0f),
+        std::make_pair(50.0f, 100.0f),
+        std::make_pair(60.0f, 100.0f),
+        std::make_pair(0.0f, 50.0f),
+        std::make_pair(-25.0f, 50.0f)
     )
 );
 
 TEST_P(FuelPercentageTest, FuelPercentage){
-    int consumedFuel = GetParam().first;
+    float consumedFuel = GetParam().first;
     float expectedPercentage = GetParam().second;
 
     FuelTank fuelTank(100);
@@ -85,11 +85,11 @@ INSTANTIATE_TEST_SUITE_P(
     FuelTankPercentageTests,
     FuelPercentageTest,
     ::testing::Values(
-        std::make_pair(90, 10.0f),
-        std::make_pair(100, 0.0f),
-        std::make_pair(150, 0.0f),
-        std::make_pair(80, 20.0f),
-        std::make_pair(0, 100.0f),
-        std::make_pair(-25, 100.0f)
+        std::make_pair(90.0f, 10.0f),
+        std::make_pair(100.0f, 0.0f),
+        std::make_pair(150.0f, 0.0f),
+        std::make_pair(80.0f, 20.0f),
+        std::make_pair(0.0f, 100.0f),
+        std::make_pair(-25.0f, 100.0f)
     )
 );
